Additional fix for text encoding
I'll be honest; I don't quite understand why the function I changed even exists in the first place. It's only called by a function in ZipCrypto, which can easily be changed to just call the function above it (since that's all it does itself). But whatever, it's a rant I'll save for another day.