Tweak properties in the netstandard project file
This makes the properties consistent with those projects which are
currently packaged as NuGet files.  Those projects are specifically the
the main package, the Android package, and the iOS package.

The downside is we become less consistent with, for example, the
Reduced DLL.  But it turns out Reduced has not had a NuGet package
uploaded for several years.