Upgrade to Albacore v3
This will allow us to create a NuGet package of the existing .NET
Standard DLL.

Upgrading Albacore requires using Paket to generate the NuGet packages,
instead of directly using NuGet.exe.  Paket can conveniently be
installed as a .NET Core Global Tool; this means we do not need to
check into source control the binary paket.exe, nor
paket.bootstrapper.exe.

Upgrading Albacore also both allows and requires specifying the NuGet
package ID in the project file.  This is instead of in the Albacore
metadata object.  Otherwise, Albacore defaults the ID to the assembly
name, the latter of which is the same across all our different
packages.  In hindsight, being able to specify the package ID in the
project file is quite nice, because the ID would also be picked up by
other tools such as NuGet.exe.