Change default iOS target platform to AnyCPU
The old default, as specified in the project file, is iPhoneSimulator.
Regardless of this change, building the NuGet package via Rake
correctly instead targets AnyCPU.  That is, something correctly
overrides the default Platform; this happens presumably by explicitly
setting the Platform property on the msbuild command line.

That means this change doesn't immediately affect our NuGet package
builds.  So why make the change?  Because when we upgrade to Albacore
v3, Albacore will read the default Platform from the top of the project
file.

Albacore will not know the default Platform is overriden to be AnyCPU.
When determining the OutputPath, Albacore will try to look up the
PropertyGroup for the default Platform, because Albacore is unaware of
any Platform property set on the msbuild command line.  This means,
unless we change the default Platform, Albacore will end up with the
wrong OutputPath.  This would then prevent a NuGet package, of the
outputted DLL, being created.